Removes the unnecessary for loop on file write
Removes the double nested for loop that iterated over z and x axis, respectively, and dumped x axis from the one dimension array to the file, since the data is already in the right order to be dumped directly/entirely to the file. This saves significant time for the dumping to the file.